优化权限校验

+++ web-project1/src/main/java/com/itcast/service/impl/ClazzServiceImpl.java
@@ -43,6 +43,10 @@ public class ClazzServiceImpl implements ClazzService {
 
     @Override
     public void delete(Integer id) {
+        long count = clazzMapper.countStudentByClazzId(id);
+        if (count > 0) {
+            throw new RuntimeException("该班级下有学生，不能删除");
+        }
         clazzMapper.delete(id);
     }
